+++ R/rUM-package.R
@@ -2,5 +2,17 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom tidyverse tidyverse_logo
+#' @importFrom conflicted conflict_prefer
+#' @importFrom bookdown html_document2
+#' @importFrom rmarkdown html_document
+#' @importFrom table1 t1kable
+#' @importFrom rlang abort
+#' @importFrom utils download.file
+#' @importFrom usethis create_project ui_done ui_info ui_nope ui_oops
+#' @importFrom glue glue
+#' @importFrom gtsummary tbl_summary
+#' @importFrom rio import
+#' @importFrom tidymodels tidymodels_prefer
 ## usethis namespace: end
 NULL